requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
transformers==4.38.2
torch==2.2.1
pillow==10.2.0
requests==2.31.0
omegaconf==2.3.0
python-multipart==0.0.9